PP-4508 Renamed TransactionType to TransactionSearchStrategyTransactionType
- The name `TransactionType` is too generic and we want to introduce another
  enum `TransactionType` which will have `CHARGE` instead of `PAYMENT`.

with @alexbishop1 and @stephencdaly
