+++ src/main/java/uk/gov/pay/connector/charge/model/TransactionSearchStrategyTransactionType.java
@@ -2,28 +2,28 @@ package uk.gov.pay.connector.charge.model;
 
 import java.util.List;
 
-public enum TransactionType {
+public enum TransactionSearchStrategyTransactionType {
     PAYMENT("payment"), REFUND("refund");
 
     private String value;
 
-    TransactionType(String value) {
+    TransactionSearchStrategyTransactionType(String value) {
         this.value = value;
     }
 
-    public static TransactionType inferTransactionTypeFrom(List<String> paymentStates, List<String> refundStates) {
+    public static TransactionSearchStrategyTransactionType inferTransactionTypeFrom(List<String> paymentStates, List<String> refundStates) {
 
-        TransactionType transactionType = null;
+        TransactionSearchStrategyTransactionType transactionSearchStrategyTransactionType = null;
         boolean hasSpecifiedPaymentStates = paymentStates != null && !paymentStates.isEmpty();
         boolean hasSpecifiedRefundStates = refundStates != null && !refundStates.isEmpty();
 
         if (hasSpecifiedPaymentStates && !hasSpecifiedRefundStates) {
-            transactionType = PAYMENT;
+            transactionSearchStrategyTransactionType = PAYMENT;
         } else if (hasSpecifiedRefundStates && !hasSpecifiedPaymentStates) {
-            transactionType = REFUND;
+            transactionSearchStrategyTransactionType = REFUND;
         }
 
-        return transactionType;
+        return transactionSearchStrategyTransactionType;
     }
 
     public String getValue() {
