+++ src/test/java/uk/gov/pay/connector/it/dao/TransactionDaoITest.java
@@ -27,8 +27,8 @@ import static org.exparity.hamcrest.date.ZonedDateTimeMatchers.within;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.core.Is.is;
-import static uk.gov.pay.connector.charge.model.TransactionType.PAYMENT;
-import static uk.gov.pay.connector.charge.model.TransactionType.REFUND;
+import static uk.gov.pay.connector.charge.model.TransactionSearchStrategyTransactionType.PAYMENT;
+import static uk.gov.pay.connector.charge.model.TransactionSearchStrategyTransactionType.REFUND;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATION_READY;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.ENTERING_CARD_DETAILS;
 import static uk.gov.pay.connector.common.model.api.ExternalChargeState.EXTERNAL_CREATED;
